Move commit box styles to separate file
